refactor(email): extract gallery selection email template into utility
